:recycle: Move the ability to obtain DataFlavor and TransferData to the ClipAppearItem interface
