+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/item/ImagesClipItem.kt
@@ -14,8 +14,12 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import org.mongodb.kbson.BsonObjectId
 import org.mongodb.kbson.ObjectId
+import java.awt.Image
+import java.awt.datatransfer.DataFlavor
+import java.io.File
 import java.nio.file.Path
 import java.nio.file.Paths
+import javax.imageio.ImageIO
 
 @Serializable
 class ImagesClipItem: RealmObject, ClipAppearItem, ClipFiles {
@@ -79,5 +83,18 @@ class ImagesClipItem: RealmObject, ClipAppearItem, ClipFiles {
             }
         }
         realm.delete(this)
+    }
+
+    override fun fillDataFlavor(map: MutableMap<DataFlavor, Any>) {
+        val filePaths = getFilePaths()
+        if (filePaths.size == 1) {
+            val imageFile = filePaths[0].toFile()
+            val bufferedImage: Image = ImageIO.read(imageFile)
+            map[DataFlavor.imageFlavor] = bufferedImage
+            map[DataFlavor.javaFileListFlavor] = listOf(imageFile)
+        } else {
+            val fileList: List<File> = filePaths.map { it.toFile() }
+            map[DataFlavor.javaFileListFlavor] = fileList
+        }
     }
 }
