+++ composeApp/src/desktopMain/kotlin/com/clipevery/clip/item/TextClipItem.kt
@@ -9,6 +9,7 @@ import kotlinx.serialization.Serializable
 import kotlinx.serialization.Transient
 import org.mongodb.kbson.BsonObjectId
 import org.mongodb.kbson.ObjectId
+import java.awt.datatransfer.DataFlavor
 
 @Serializable
 class TextClipItem: RealmObject, ClipAppearItem, ClipText {
@@ -41,5 +42,9 @@ class TextClipItem: RealmObject, ClipAppearItem, ClipText {
 
     override fun clear(realm: MutableRealm, clearResource: Boolean) {
         realm.delete(this)
+    }
+
+    override fun fillDataFlavor(map: MutableMap<DataFlavor, Any>) {
+        map[DataFlavor.stringFlavor] = text
     }
 }
